fix(ort-scan): Allow analyzing a different revision of same repository
The logic for determining whether the downloader needs to be run in
order to analyze a repository / revision different than the already
cloned one only considers the VCS_URL. This is not sufficient. In
particular, it is not possible to scan any other revision or branch of
that same repository.

Fix that problem by the following heuristic:

1. If the user did provided at least one `VCS_*` variable, then always
   use the downloader to obtain the sources to be analyzed.
2. Otherwise, analyze the already cloned repository.

Note: In case 2, the actual sha1 is now used as VCS_REVISION instead of
"HEAD".

Signed-off-by: Frank Viernau <frank_viernau@epam.com>